# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "gaming product management with REST API",
        "description": "a scalable and flexible system for managing a store of gaming product .",
        "contact": {
            "email": "ahmed.yassin.elhamzaoui2019@gmail.com"
        },
        "license": {
            "name": "made by Ahmed Yassine El Hamzaoui",
            "url": "https://github.com/ahmedyassineelhamzaoui"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/add-product": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Add a new product",
                "description": "Adds a new product with the given details",
                "operationId": "76121ec334829c1da1ef6b8205c509b1",
                "requestBody": {
                    "description": "The product information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "price",
                                    "categorie_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 200
                                    },
                                    "description": {
                                        "type": "string",
                                        "minLength": 10
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "categorie_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product created successfully"
                                        },
                                        "product": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "title": {
                                                    "type": "string"
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "price": {
                                                    "type": "number"
                                                },
                                                "categorie_id": {
                                                    "type": "integer"
                                                },
                                                "user_id": {
                                                    "type": "integer"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Category not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Product already exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "This title of product already exists. Please enter another one."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/update-product": {
            "put": {
                "tags": [
                    "Product"
                ],
                "summary": "Update a product",
                "description": "Updates a specific product based on its ID",
                "operationId": "c3f881009b52dee85956f74770f0734f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "The ID of the product to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "The title of the product",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 200
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "The description of the product",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "minLength": 10
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "The price of the product",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "categorie_id",
                        "in": "query",
                        "description": "The ID of the category to which the product belongs",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "The status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "The message to be returned",
                                            "type": "string",
                                            "example": "product has been updated successfully"
                                        },
                                        "product": {
                                            "description": "The updated product object",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "The status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "The message to be returned",
                                            "type": "string",
                                            "example": "this id doesn't exist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/delete-product": {
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete a product",
                "description": "Deletes a specific product based on its ID",
                "operationId": "00a007c20cf0f2856c8cba8b7a01e3fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "The ID of the product to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "The status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "description": "The message to be returned",
                                            "type": "string",
                                            "example": "product deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "The status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "The message to be returned",
                                            "type": "string",
                                            "example": "the id that you enter not exist"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}